+++ rossum-agent/CHANGELOG.md
@@ -7,3 +7,4 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased] - YYYY-MM-DD
 
 - Fixed displaying mermaid diagrams in Streamlit UI [#36](https://github.com/stancld/rossum-mcp/pull/36)
+- Added beep sound notification upon completing the agent answer [#37](https://github.com/stancld/rossum-mcp/pull/37)